[to be squashed] Refactor hashtable tests.
