+++ flink-runtime/src/test/java/org/apache/flink/runtime/operators/hash/CompactingHashTableTest.java
@@ -25,25 +25,23 @@ import org.apache.flink.api.common.typeutils.TypeComparator;
 import org.apache.flink.api.common.typeutils.TypePairComparator;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.common.typeutils.base.LongComparator;
-import org.apache.flink.api.common.typeutils.base.LongSerializer;
-import org.apache.flink.api.common.typeutils.base.StringSerializer;
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.api.java.typeutils.runtime.TupleComparator;
-import org.apache.flink.api.java.typeutils.runtime.TupleSerializer;
 import org.apache.flink.core.memory.MemorySegment;
 import org.apache.flink.core.memory.MemorySegmentFactory;
+import org.apache.flink.runtime.operators.testutils.types.IntList;
+import org.apache.flink.runtime.operators.testutils.types.IntPair;
 import org.apache.flink.util.MutableObjectIterator;
 
 import org.junit.Test;
+import org.powermock.reflect.Whitebox;
 
+import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.List;
+import java.util.Random;
 
-public class CompactingHashTableTest {
-
-	private final TypeSerializer<Tuple2<Long, String>> serializer;
-	private final TypeComparator<Tuple2<Long, String>> comparator;
+public class CompactingHashTableTest extends MutableHashTableTestBase {
 	
 	private final TypeComparator<Long> probeComparator;
 
@@ -51,16 +49,7 @@ public class CompactingHashTableTest {
 	
 	
 	public CompactingHashTableTest() {
-		TypeSerializer<?>[] fieldSerializers = { LongSerializer.INSTANCE, StringSerializer.INSTANCE };
-		@SuppressWarnings("unchecked")
-		Class<Tuple2<Long, String>> clazz = (Class<Tuple2<Long, String>>) (Class<?>) Tuple2.class;
-		this.serializer = new TupleSerializer<Tuple2<Long, String>>(clazz, fieldSerializers);
-		
-		TypeComparator<?>[] comparators = { new LongComparator(true) };
-		TypeSerializer<?>[] comparatorSerializers = { LongSerializer.INSTANCE };
-		
-		this.comparator = new TupleComparator<Tuple2<Long, String>>(new int[] {0}, comparators, comparatorSerializers);
-		
+
 		this.probeComparator = new LongComparator(true);
 		
 		this.pairComparator = new TypePairComparator<Long, Tuple2<Long, String>>() {
@@ -86,11 +75,20 @@ public class CompactingHashTableTest {
 			}
 		};
 	}
-	
+
+	@Override
+	protected <T> AbstractMutableHashTable<T> getHashTable(TypeSerializer<T> serializer, TypeComparator<T> comparator, List<MemorySegment> memory) {
+		return new CompactingHashTable<T>(serializer, comparator, memory);
+	}
+
 	// ------------------------------------------------------------------------
 	//  tests
 	// ------------------------------------------------------------------------
-	
+
+	/**
+	 * This has to be duplicated in InPlaceMutableHashTableTest and CompactingHashTableTest
+	 * because of the different constructor calls.
+	 */
 	@Test
 	public void testHashTableGrowthWithInsert() {
 		try {
@@ -102,7 +100,7 @@ public class CompactingHashTableTest {
 			// a lot of memory for the partition buffers, and start with a smaller hash table. that way
 			// we trigger a hash table growth early.
 			CompactingHashTable<Tuple2<Long, String>> table = new CompactingHashTable<Tuple2<Long, String>>(
-					serializer, comparator, memory, 10000);
+				tuple2LongStringSerializer, tuple2LongStringComparator, memory, 10000);
 			table.open();
 			
 			for (long i = 0; i < numElements; i++) {
@@ -144,6 +142,9 @@ public class CompactingHashTableTest {
 
 	/**
 	 * This test validates that records are not lost via "insertOrReplace()" as in bug [FLINK-2361]
+	 *
+	 * This has to be duplicated in InPlaceMutableHashTableTest and CompactingHashTableTest
+	 * because of the different constructor calls.
 	 */
 	@Test
 	public void testHashTableGrowthWithInsertOrReplace() {
@@ -155,8 +156,8 @@ public class CompactingHashTableTest {
 			// we create a hash table that thinks the records are super large. that makes it choose initially
 			// a lot of memory for the partition buffers, and start with a smaller hash table. that way
 			// we trigger a hash table growth early.
-			CompactingHashTable<Tuple2<Long, String>> table = new CompactingHashTable<Tuple2<Long, String>>(
-					serializer, comparator, memory, 10000);
+			CompactingHashTable<Tuple2<Long, String>> table = new CompactingHashTable<>(
+				tuple2LongStringSerializer, tuple2LongStringComparator, memory, 10000);
 			table.open();
 			
 			for (long i = 0; i < numElements; i++) {
@@ -212,8 +213,8 @@ public class CompactingHashTableTest {
 			// we create a hash table that thinks the records are super large. that makes it choose initially
 			// a lot of memory for the partition buffers, and start with a smaller hash table. that way
 			// we trigger a hash table growth early.
-			CompactingHashTable<Tuple2<Long, String>> table = new CompactingHashTable<Tuple2<Long, String>>(
-					serializer, comparator, memory, 100);
+			CompactingHashTable<Tuple2<Long, String>> table = new CompactingHashTable<>(
+				tuple2LongStringSerializer, tuple2LongStringComparator, memory, 100);
 			table.open();
 
 			// first, we insert some elements
@@ -234,7 +235,7 @@ public class CompactingHashTableTest {
 
 			// check the results
 			CompactingHashTable<Tuple2<Long, String>>.HashTableProber<Tuple2<Long, String>> prober =
-				table.getProber(comparator, new SameTypePairComparator<>(comparator));
+				table.getProber(tuple2LongStringComparator, new SameTypePairComparator<>(tuple2LongStringComparator));
 			Tuple2<Long, String> reuse = new Tuple2<>();
 			for (long i = 0; i < numElements; i++) {
 				assertNotNull(prober.getMatchFor(Tuple2.of(i, longString), reuse));
@@ -244,6 +245,233 @@ public class CompactingHashTableTest {
 		catch (Exception e) {
 			e.printStackTrace();
 			fail(e.getMessage());
+		}
+	}
+
+	@Test
+	public void testResize() {
+		// Only CompactingHashTable
+		try {
+			final int NUM_MEM_PAGES = 30 * NUM_PAIRS / PAGE_SIZE;
+			final Random rnd = new Random(RANDOM_SEED);
+			final IntPair[] pairs = getRandomizedIntPairs(NUM_PAIRS, rnd);
+
+			List<MemorySegment> memory = getMemory(NUM_MEM_PAGES);
+			CompactingHashTable<IntPair> table = new CompactingHashTable<IntPair>(intPairSerializer, intPairComparator, memory);
+			table.open();
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				table.insert(pairs[i]);
+			}
+
+			AbstractHashTableProber<IntPair, IntPair> prober =
+				table.getProber(intPairComparator, new SameTypePairComparator<>(intPairComparator));
+			IntPair target = new IntPair();
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(ADDITIONAL_MEM));
+			Boolean b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(pairs[i].getKey() + " " + pairs[i].getValue(), prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			table.close();
+			assertEquals("Memory lost", NUM_MEM_PAGES + ADDITIONAL_MEM, table.getFreeMemory().size());
+		} catch (Exception e) {
+			e.printStackTrace();
+			fail("Error: " + e.getMessage());
+		}
+	}
+
+	@Test
+	public void testDoubleResize() {
+		// Only CompactingHashTable
+		try {
+			final int NUM_MEM_PAGES = 30 * NUM_PAIRS / PAGE_SIZE;
+			final Random rnd = new Random(RANDOM_SEED);
+			final IntPair[] pairs = getRandomizedIntPairs(NUM_PAIRS, rnd);
+
+			List<MemorySegment> memory = getMemory(NUM_MEM_PAGES);
+			CompactingHashTable<IntPair> table = new CompactingHashTable<IntPair>(intPairSerializer, intPairComparator, memory);
+			table.open();
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				table.insert(pairs[i]);
+			}
+
+			AbstractHashTableProber<IntPair, IntPair> prober =
+				table.getProber(intPairComparator, new SameTypePairComparator<>(intPairComparator));
+			IntPair target = new IntPair();
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(ADDITIONAL_MEM));
+			Boolean b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(pairs[i].getKey() + " " + pairs[i].getValue(), prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(ADDITIONAL_MEM));
+			b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(pairs[i].getKey() + " " + pairs[i].getValue(), prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			table.close();
+			assertEquals("Memory lost", NUM_MEM_PAGES + ADDITIONAL_MEM + ADDITIONAL_MEM, table.getFreeMemory().size());
+		} catch (Exception e) {
+			e.printStackTrace();
+			fail("Error: " + e.getMessage());
+		}
+	}
+
+	@Test
+	public void testTripleResize() {
+		// Only CompactingHashTable
+		try {
+			final int NUM_MEM_PAGES = 30 * NUM_PAIRS / PAGE_SIZE;
+			final Random rnd = new Random(RANDOM_SEED);
+			final IntPair[] pairs = getRandomizedIntPairs(NUM_PAIRS, rnd);
+
+			List<MemorySegment> memory = getMemory(NUM_MEM_PAGES);
+			CompactingHashTable<IntPair> table = new CompactingHashTable<IntPair>(intPairSerializer, intPairComparator, memory);
+			table.open();
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				table.insert(pairs[i]);
+			}
+
+			AbstractHashTableProber<IntPair, IntPair> prober =
+				table.getProber(intPairComparator, new SameTypePairComparator<>(intPairComparator));
+			IntPair target = new IntPair();
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(ADDITIONAL_MEM));
+			Boolean b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(pairs[i].getKey() + " " + pairs[i].getValue(), prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(ADDITIONAL_MEM));
+			b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(pairs[i].getKey() + " " + pairs[i].getValue(), prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(2*ADDITIONAL_MEM));
+			b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_PAIRS; i++) {
+				assertNotNull(pairs[i].getKey() + " " + pairs[i].getValue(), prober.getMatchFor(pairs[i], target));
+				assertEquals(pairs[i].getValue(), target.getValue());
+			}
+
+			table.close();
+			assertEquals("Memory lost", NUM_MEM_PAGES + 4*ADDITIONAL_MEM, table.getFreeMemory().size());
+		} catch (Exception e) {
+			e.printStackTrace();
+			fail("Error: " + e.getMessage());
+		}
+	}
+
+	@Test
+	public void testResizeWithCompaction(){
+		// Only CompactingHashTable
+		try {
+			final int NUM_MEM_PAGES = (SIZE * NUM_LISTS / PAGE_SIZE);
+
+			final Random rnd = new Random(RANDOM_SEED);
+			final IntList[] lists = getRandomizedIntLists(NUM_LISTS, rnd);
+
+			List<MemorySegment> memory = getMemory(NUM_MEM_PAGES);
+			CompactingHashTable<IntList> table = new CompactingHashTable<IntList>(serializerV, comparatorV, memory);
+			table.open();
+
+			for (int i = 0; i < NUM_LISTS; i++) {
+				table.insert(lists[i]);
+			}
+
+			AbstractHashTableProber<IntList, IntList> prober = table.getProber(comparatorV, pairComparatorV);
+			IntList target = new IntList();
+
+			for (int i = 0; i < NUM_LISTS; i++) {
+				assertNotNull(prober.getMatchFor(lists[i], target));
+				assertArrayEquals(lists[i].getValue(), target.getValue());
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(ADDITIONAL_MEM));
+			Boolean b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_LISTS; i++) {
+				assertNotNull(prober.getMatchFor(lists[i], target));
+				assertArrayEquals(lists[i].getValue(), target.getValue());
+			}
+
+			final IntList[] overwriteLists = getRandomizedIntLists(NUM_LISTS, rnd);
+
+			// test replacing
+			for (int i = 0; i < NUM_LISTS; i++) {
+				table.insertOrReplaceRecord(overwriteLists[i]);
+			}
+
+			Field list = Whitebox.getField(CompactingHashTable.class, "partitions");
+			@SuppressWarnings("unchecked")
+			ArrayList<InMemoryPartition<IntList>> partitions = (ArrayList<InMemoryPartition<IntList>>) list.get(table);
+			int numPartitions = partitions.size();
+			for(int i = 0; i < numPartitions; i++) {
+				Whitebox.invokeMethod(table, "compactPartition", i);
+			}
+
+			// make sure there is enough memory for resize
+			memory.addAll(getMemory(2*ADDITIONAL_MEM));
+			b = Whitebox.<Boolean>invokeMethod(table, "resizeHashTable");
+			assertTrue(b);
+
+			for (int i = 0; i < NUM_LISTS; i++) {
+				assertNotNull("" + i, prober.getMatchFor(overwriteLists[i], target));
+				assertArrayEquals(overwriteLists[i].getValue(), target.getValue());
+			}
+
+			table.close();
+			assertEquals("Memory lost", NUM_MEM_PAGES + 3*ADDITIONAL_MEM, table.getFreeMemory().size());
+		} catch (Exception e) {
+			e.printStackTrace();
+			fail("Error: " + e.getMessage());
 		}
 	}
 	
@@ -261,82 +489,5 @@ public class CompactingHashTableTest {
 			bld.append('a');
 		}
 		return bld.toString();
-	}
-
-
-
-
-
-
-	@Test
-	public void testDifferentProbers() {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = 32 * common.NUM_PAIRS / common.PAGE_SIZE;
-		common.testDifferentProbers(new CompactingHashTable<>(common.serializer, common.comparator, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testBuildAndRetrieve() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = 32 * common.NUM_PAIRS / common.PAGE_SIZE;
-		common.testBuildAndRetrieve(new CompactingHashTable<>(common.serializer, common.comparator, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testEntryIterator() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testEntryIterator(new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testMultipleProbers() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testMultipleProbers(new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testVariableLengthBuildAndRetrieve() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testVariableLengthBuildAndRetrieve(new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testVariableLengthBuildAndRetrieveMajorityUpdated() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testVariableLengthBuildAndRetrieveMajorityUpdated(new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testVariableLengthBuildAndRetrieveMinorityUpdated() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_LISTS = 20000;
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testVariableLengthBuildAndRetrieveMinorityUpdated(
-			new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES, NUM_LISTS);
-	}
-
-	@Test
-	public void testRepeatedBuildAndRetrieve() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testRepeatedBuildAndRetrieve(new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testProberUpdate() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = common.SIZE * common.NUM_LISTS / common.PAGE_SIZE;
-		common.testProberUpdate(new CompactingHashTable<>(common.serializerV, common.comparatorV, common.getMemory(NUM_MEM_PAGES)), NUM_MEM_PAGES);
-	}
-
-	@Test
-	public void testVariableLengthStringBuildAndRetrieve() throws Exception {
-		MemoryHashTableTestCommon common = new MemoryHashTableTestCommon();
-		final int NUM_MEM_PAGES = 40 * common.NUM_PAIRS / common.PAGE_SIZE;
-		common.testVariableLengthStringBuildAndRetrieve(NUM_MEM_PAGES, new CompactingHashTable<>(common.serializerS, common.comparatorS, common.getMemory(NUM_MEM_PAGES)));
 	}
 }
